requirements.txt
torch
transformers
torchvision
sentencepiece==0.1.99
pytesseract==0.3.10
Pillow==9.5.0
scipy==1.10.1
opencv-python==4.7.0.72
psycopg2-binary==2.9.6  # PostgreSQL 驅動